Inline ToMessage and ToBytes to the ServiceStackTextMessageByteSerializer
